[front] chore: move MCP server view helper out of models
