+++ front/lib/resources/mcp_server_view_helper.ts
